pyproject.toml
[project]
name = "uv-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.4",
    "pytesseract>=0.3.13",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "tools>=1.0.5",
    "typing-extensions>=4.15.0",
    "volcengine-python-sdk>=4.0.14",
]
